+++ client/order/src/main/java/com/live_commerce/order/application/dto/response/OrderUpdateResponse.java
@@ -11,7 +11,7 @@ public record OrderUpdateResponse(
         UUID orderId,
         UUID productId,
         Integer productQuantity,
-        Long productTotalPrice,
+        Double productTotalPrice,
         String requirement,
         String status,
         LocalDateTime updatedAt
